improving cra
