+++ examples/example-app-cra/src/App.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import logo from './logo.svg';
 import './App.css';
+import {What} from "./what";
 
 function App() {
   return (
@@ -8,7 +9,7 @@ function App() {
       <header className="App-header">
         <img src={logo} className="App-logo" alt="logo" />
         <p>
-          Edit <code>src/App.js</code> and save to reload.
+          Edit this <code>src/App.js</code> and save to reload.
         </p>
         <a
           className="App-link"
@@ -18,6 +19,7 @@ function App() {
         >
           Learn React
         </a>
+        <What/>
       </header>
     </div>
   );
