+++ examples/example-app-cra/src/what.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export function What(){
+    return <div>Hello from typescript</div>
+}
